feat(shared): added navigation menu component
