+++ src/shared/layouts/favorites-layout/favorites-layout.module.scss
@@ -8,27 +8,6 @@
     gap: 24px;
 }
 
-.linksContainer {
-    display: flex;
-}
-.link {
-    display: block;
-    padding: 8px 16px;
-    text-decoration: none;
-    align-items: center;
-    font-family: $--open-sans-font;
-    font-weight: 700;
-    line-height: normal;
-    letter-spacing: 0.25px;
-    font-size: 16px;
-    color: $--Foreground-subtitle
-}
-.linkActive {
-    border-bottom: 3px solid $--Foreground-default;
-    color: $--Foreground-default;
-}
-
-
 @media screen and (max-width: 1199px) {
     .favorites {
         width: 780px;
