+++ src/shared/layouts/favorites-layout/favorites-layout.tsx
@@ -1,9 +1,7 @@
 'use client';
 import React from 'react';
-import Link from 'next/link';
-import clsx from 'clsx';
-import { usePathname } from 'next/navigation';
 import { ProfileLink } from '@/shared/ui/profile-link/profile-link';
+import { NavigationMenu } from '@/shared/ui';
 import styles from './favorites-layout.module.scss';
 
 export const FavoritesLayout = ({
@@ -11,31 +9,16 @@ export const FavoritesLayout = ({
 }: {
 	children: React.ReactNode;
 }) => {
-	const links = [
-		{ name: 'Специалисты', href: '/favorites/specialists' },
-		{ name: 'Проекты', href: '/favorites/projects' },
-	];
-	const pathname = usePathname();
 	return (
 		<>
-		<ProfileLink title='Избранные'/>
+			<ProfileLink title="Избранные" />
 			<div className={styles.favorites}>
-				<div className={styles.linksContainer}>
-					{links.map((link) => {
-						return (
-							<Link
-								key={link.name}
-								href={link.href}
-								className={clsx(styles.link, {
-									[styles.linkActive]: pathname === link.href,
-								})}>
-								{link.name}
-							</Link>
-						);
-					})}
-				</div>
-         
-				
+				<NavigationMenu
+					linksArray={[
+						{ name: 'Специалисты', href: '/favorites/specialists' },
+						{ name: 'Проекты', href: '/favorites/projects' },
+					]}
+				/>
 				{children}
 			</div>
 		</>
